change value_max_bytes value

+++ app/server.rb
@@ -71,7 +71,7 @@ module WhosGotDirt
             connection.response :caching do
               ActiveSupport::Cache::MemCacheStore.new(ENV['MEMCACHIER_SERVERS'], {
                 expires_in: 2600, # 1 hour
-                value_max_bytes: Integer(1048576), # 1 MB
+                value_max_bytes: Integer(2097152), # 2 MB, OpenDuka can respond with >1 MB
                 username: ENV['MEMCACHIER_USERNAME'],
                 password: ENV['MEMCACHIER_PASSWORD'],
               })
